refactor(actionsPanel): extract search logic into a custom hook
